Eliminate format-string-related warnings in pe, macho, dex modules

+++ libyara/modules/macho.c
@@ -97,7 +97,7 @@ bool macho_rva_to_offset(
     YR_OBJECT* object)
 {
   uint64_t segment_count = get_integer(object, "number_of_segments");
-  for (uint64_t i = 0; i < segment_count; i++)
+  for (int i = 0; i < segment_count; i++)
   {
     uint64_t start = get_integer(object, "segments[%i].vmaddr", i);
     uint64_t end = start + get_integer(object, "segments[%i].vmsize", i);
@@ -121,7 +121,7 @@ int macho_offset_to_rva(
     YR_OBJECT* object)
 {
   uint64_t segment_count = get_integer(object, "number_of_segments");
-  for (uint64_t i = 0; i < segment_count; i++)
+  for (int i = 0; i < segment_count; i++)
   {
     uint64_t start = get_integer(object, "segments[%i].fileoff", i);
     uint64_t end = start + get_integer(object, "segments[%i].filesize", i);
@@ -722,7 +722,7 @@ define_function(file_index_type)
   if (is_undefined(module, "nfat_arch"))
     return_integer(UNDEFINED);
 
-  for (uint64_t i = 0; i < nfat; i++)
+  for (int i = 0; i < nfat; i++)
   {
     int64_t type = get_integer(module, "file[%i].cputype", i);
     if (type == type_arg)
@@ -746,7 +746,7 @@ define_function(file_index_subtype)
   if (is_undefined(module, "nfat_arch"))
     return_integer(UNDEFINED);
 
-  for (uint64_t i = 0; i < nfat; i++)
+  for (int i = 0; i < nfat; i++)
   {
     int64_t type = get_integer(module, "file[%i].cputype", i);
     int64_t subtype = get_integer(module, "file[%i].cpusubtype", i);
@@ -771,7 +771,7 @@ define_function(ep_for_arch_type)
   if (is_undefined(module, "nfat_arch"))
     return_integer(UNDEFINED);
 
-  for (uint64_t i = 0; i < nfat; i++)
+  for (int i = 0; i < nfat; i++)
   {
     int64_t type = get_integer(module, "fat_arch[%i].cputype", i);
     if (type == type_arg)
@@ -797,7 +797,7 @@ define_function(ep_for_arch_subtype)
   if (is_undefined(module, "nfat_arch"))
     return_integer(UNDEFINED);
 
-  for (uint64_t i = 0; i < nfat; i++)
+  for (int i = 0; i < nfat; i++)
   {
     int64_t type = get_integer(module, "fat_arch[%i].cputype", i);
     int64_t subtype = get_integer(module, "fat_arch[%i].cpusubtype", i);
